Já diminui a BoardModel, criei uma classe para as cenas de criar o ficheiro, e criei outro para a logica das bolas de neve
